// sourcecode/BTrak.Packages/snova-project-management/PM-module-components/pm-components/src/lib/snovasys-projects/models/workflowStatus.ts
export class WorkflowStatus {
  workflowStatusName: string;
  workFlowId: string;
  workFlowStatusId: string;
  statusId: any[];
  orderId: number;
  maxOrder: number;
  isComplete: boolean;
  isArchived: boolean;
  existingUserStoryStatusId: string;
  currentUserStoryStatusId: string;
  userStoryStatusId: string;
  taskStatusId: string;
  userStoryStatusName: string;
  timeStamp: any;
  workflowName: string;
  canAdd: any;
}