// sourcecode/BTrak.Packages/snova-project-management/PM-module-components/pm-components/src/lib/snovasys-projects/models/workFlowStatusTransitionTableData.ts
export class WorkFlowStatusTransitionTableData {
  fromTransition: string;
  toTransition: string;
  deadline: string;
  roles: any[];
  workflowEligibleStatusTransitionId: string;
  workFlowId: string;
  currentTransition: string;
  isArchived: boolean;
  fromWorkflowUserStoryStatusId: string;
  toWorkflowUserStoryStatusId: string;
  fromWorkflowUserStoryStatus: string;
  toWorkflowUserStoryStatus: string;
  roleIds: string;
  transitionDeadlineId: string;
  projectId: string;
  roleGuids: any[];
  goalId: string;
  timeStamp:any;
  sprintId: string;
}